requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
braceexpand==0.1.7
chardet==4.0.0
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
discord==1.7.3
discord.py==1.7.3
idna==3.3
import-expression==1.1.4
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
Pygments==2.11.2
PyYAML==6.0
rich==11.1.0
six==1.16.0
tomli==2.0.0
typing-extensions==4.0.1
yarl==1.7.2
